+++ stubBrowser.js
